io: move hashing code to separate file
